+++ CHANGES.rst
@@ -14,7 +14,10 @@ It is currently undergoing fast development and backward compatibility is not en
 
 Major changes
 -------------
-* The :class:`SentenceEncoder` is now available to encode columns with diverse entries.
+* The :class:`TextEncoder` is now available to encode columns with diverse entries.
+  It allows the representation of table entries as embeddings computed by a deep
+  learning model. The weights of this model can be fetched locally or from the
+  HuggingFace Hub.
   :pr:`1077` by :user:`Vincent Maladiere <Vincent-Maladiere>`.
 
 Minor changes